Assign, don't retain delegate.
I have a controller that was storing a TwitterTweetPoster object and assigning
itself as a delegate to the TwitterTweetPoster. However, because the
TwitterTweetPoster was retaining its delegate, I ended up with a circular
retain. The result was that the controller was never released from memory, and
neither was the TwitterTweetPoster.

Standard practice in iOS apps is to make delegate properties non-retaining.
Looks like all the other delegates in this package use `assign`; it was just
this one that did not.